Hooks uses jsx files and handleLogout inside hooks - renamed
